Fix: Absolute path bug (#6)
* Fix: Absolute path bug

References can look like this: [sometext](/docs/test.md).

Problem: The reference /docs/test.md isn't actually an absolute reference, but a reference related to some other parent folder that is being used as a root folder.

Therefore, references with this syntax have to be tested twice:

The existence of the file as is, which means as an absolute file path
The existence of the file related to some parent directory.
To solve 2., you would have to take the absolute path of the file where this reference was made in and traverse it upwards, testing in each upwards movement step, if the file exists relative to that path.